♻️ refactor: 캐시 삭제 로직 수정
