+++ src/components/common/Dropdown.tsx
@@ -7,7 +7,6 @@ import useUserStore from '@/zustand/useStore';
 import Swal from 'sweetalert2';
 // import { signOut } from 'next-auth/react';
 import { usePathname, useRouter } from 'next/navigation';
-import { Logout } from '@/data/actions/user';
 
 export default function Dropdown() {
   const { resetUser } = useUserStore();
@@ -20,12 +19,16 @@ export default function Dropdown() {
     pathname === path ? 'mypage-dropdown-active' : '';
 
   //로그아웃 시 토큰 삭제
-  const handleLogout = () => {
+  const handleLogout = async () => {
     resetUser();
     localStorage.removeItem('accessToken');
     localStorage.removeItem('refreshToken');
     localStorage.removeItem('userInfo');
-    Logout();
+
+    await fetch('/api/logout', {
+      method: 'POST',
+      credentials: 'include', // 쿠키 포함 요청 (필요 시)
+    });
 
     Swal.fire({
       icon: 'info',
