+++ src/data/actions/user.ts
@@ -3,8 +3,6 @@
 import { signIn } from '@/auth';
 import { ApiRes, ApiResPromise, User } from '@/types';
 import { OAuthUser } from '@/types';
-import { cookies } from 'next/headers';
-import { NextResponse } from 'next/server';
 const API_URL = process.env.NEXT_PUBLIC_API_URL;
 const CLIENT_ID = process.env.NEXT_PUBLIC_CLIENT_ID || '';
 
@@ -184,29 +182,4 @@ export async function loginWithAuthjs(provider: string, formData: FormData) {
   const redirectTo = (formData.get('redirect') as string) || '/';
 
   await signIn(provider, { redirectTo });
-}
-
-// 로그아웃
-export async function Logout() {
-  (await cookies()).delete('authjs.session-token');
-  (await cookies()).delete('authjs.callback-url');
-
-  // 배포용 쿠키 삭제
-  const response = NextResponse.json({ ok: true });
-
-  response.cookies.set('__Secure-authjs.session-token', '', {
-    maxAge: 0,
-    path: '/',
-    httpOnly: true,
-    secure: true,
-    sameSite: 'lax',
-  });
-
-  response.cookies.set('__Secure-authjs.callback-url', '', {
-    maxAge: 0,
-    path: '/',
-    httpOnly: true,
-    secure: true,
-    sameSite: 'lax',
-  });
 }
